change various integer types to NSUInteger, per warnings

+++ src/Action.h
@@ -25,7 +25,7 @@
 - (ImageProcessor *)keyedImageProcessor;
 - (NSString *)keyParameter;
 - (NSString *)parameterAtIndex:(int)i;
-- (int)parameterCount;
+- (NSUInteger)parameterCount;
 - (ImageProcessor *)imageProcessorForKey:(NSString *)key;
 
 @end
